fix(engine): make a ps status filter imply querying every container
`ps_filtered` queried libpod with `all=${opts.all}` even when a
`--status`/`--filter status=` predicate was set. libpod's list endpoint
only returns running containers when `all=false`, so `podup ps
--status exited` (without `-a`) always came back empty, even with
matching exited containers present. I now OR the status filter into the
query's `all` flag, so a status filter widens what libpod is asked for;
the existing client-side `status_matches` filtering still narrows the
result to what was actually requested.

Added two tests against the fake libpod harness asserting the outgoing
request's `all=` value: forced to `true` when a status filter is set
with `opts.all=false`, and left at `false` with no filter. This pushed
ps.rs past the 500-line hard limit, so I split its test module into a
sibling `ps/tests.rs`, matching the pattern already used for lifecycle.
